Add support for toggling "eco mode" on and off
On units with the old-style EDA control panel the mode doesn't seem to be reflected on the control panel, but the Modbus register sticks and the update seems to go through. Would have to monitor the device to see whether target temperature is adjusted accordingly when eco mode is on.